split off common and server (remote) functions
