+++ include/ctp7_modules/common/vfat3.h
@@ -5,26 +5,18 @@
  *  \author Brian Dorney <brian.l.dorney@cern.ch>
  */
 
-#ifndef VFAT3_H
-#define VFAT3_H
+#ifndef COMMON_VFAT3_H
+#define COMMON_VFAT3_H
 
-#include "utils.h"
-#include "hw_constants.h"
+#include "ctp7_modules/common/hw_constants.h"
+
+#include "xhal/common/rpc/common.h"
 
 #include <map>
 #include <string>
 #include <vector>
 
 namespace vfat3 {
-
-  /*!
-   *  \brief Decode a Reed--Muller encoded VFAT3 ChipID
-   *
-   *  \param \c encChipID 32-bit encoded chip ID to decode
-   *
-   *  \returns decoded VFAT3 chip ID
-   */
-  uint16_t decodeChipID(const uint32_t &encChipID);
 
   /*!
    *  \brief Verify whether VFATs on a specified front end optical link are in sync
