+++ src/amc.cpp
@@ -6,11 +6,14 @@
  */
 
 #include "ctp7_modules/common/amc.h"
+#include "ctp7_modules/server/amc.h"
 #include "ctp7_modules/common/amc/ttc.h"
 #include "ctp7_modules/common/amc/daq.h"
 #include "ctp7_modules/common/amc/sca.h"
 #include "ctp7_modules/common/amc/blaster_ram.h"
 #include "ctp7_modules/common/hw_constants.h"
+#include "ctp7_modules/common/utils.h"
+#include "ctp7_modules/server/utils.h"
 
 #include "xhal/common/rpc/register.h"
