+++ src/amc/ttc.cpp
@@ -4,6 +4,8 @@
  */
 
 #include "ctp7_modules/common/amc/ttc.h"
+#include "ctp7_modules/common/utils.h"
+#include "ctp7_modules/server/utils.h"
 
 #include <chrono>
 #include <iomanip>
