+++ src/utils.cpp
@@ -1,4 +1,5 @@
 #include "ctp7_modules/common/utils.h"
+#include "ctp7_modules/server/utils.h"
 
 #include <log4cplus/configurator.h>
 #include <log4cplus/hierarchy.h>
